Move some common test code to the tools:unit package
