+++ dd-sdk-android/src/test/kotlin/com/datadog/android/utils/forge/Configurator.kt
@@ -63,11 +63,6 @@ internal class Configurator :
         // MISC
         forge.addFactory(BigIntegerFactory())
         forge.addFactory(CharsetForgeryFactory())
-        forge.addFactory(GsonJsonArrayForgeryFactory())
-        forge.addFactory(GsonJsonObjectForgeryFactory())
-        forge.addFactory(GsonJsonPrimitiveForgeryFactory())
-        forge.addFactory(OrgJSONObjectForgeryFactory())
-        forge.addFactory(OrgJSONArrayForgeryFactory())
 
         // Datadog SDK v2
         forge.addFactory(TimeInfoForgeryFactory())
