+++ dd-sdk-android/src/test/kotlin/com/datadog/android/utils/forge/LongTaskEventForgeryFactory.kt
@@ -9,6 +9,7 @@ package com.datadog.android.utils.forge
 import com.datadog.android.core.internal.system.AndroidInfoProvider
 import com.datadog.android.rum.internal.domain.scope.toLongTaskSchemaType
 import com.datadog.android.rum.model.LongTaskEvent
+import com.datadog.tools.unit.forge.exhaustiveAttributes
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
 import fr.xgouchet.elmyr.jvm.ext.aTimestamp
