+++ library/dd-sdk-android-logs/src/test/kotlin/com/datadog/android/utils/forge/Configurator.kt
@@ -20,13 +20,6 @@ internal class Configurator : BaseConfigurator() {
         forge.addFactory(UserInfoForgeryFactory())
         forge.addFactory(NetworkInfoForgeryFactory())
 
-        forge.addFactory(GsonJsonArrayForgeryFactory())
-        forge.addFactory(GsonJsonObjectForgeryFactory())
-        forge.addFactory(GsonJsonPrimitiveForgeryFactory())
-
-        forge.addFactory(OrgJSONArrayForgeryFactory())
-        forge.addFactory(OrgJSONObjectForgeryFactory())
-
         forge.useJvmFactories()
     }
 }
