+++ library/dd-sdk-android-logs/src/test/kotlin/com/datadog/android/utils/forge/DatadogContextForgeryFactory.kt
@@ -13,6 +13,7 @@ import com.datadog.android.v2.api.context.DeviceInfo
 import com.datadog.android.v2.api.context.DeviceType
 import com.datadog.android.v2.api.context.ProcessInfo
 import com.datadog.android.v2.api.context.TimeInfo
+import com.datadog.tools.unit.forge.exhaustiveAttributes
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
 import java.util.Locale
@@ -67,9 +68,7 @@ internal class DatadogContextForgeryFactory : ForgeryFactory<DatadogContext> {
             // building nested maps with default size slows down tests quite a lot, so will use
             // an explicit small size
             featuresContext = forge.aMap(size = 2) {
-                forge.anAlphabeticalString() to forge.aMap {
-                    forge.anAlphabeticalString() to forge.anAlphabeticalString()
-                }
+                forge.anAlphabeticalString() to forge.exhaustiveAttributes()
             }
         )
     }
