+++ library/dd-sdk-android-logs/src/test/kotlin/com/datadog/android/utils/forge/UserInfoForgeryFactory.kt
@@ -7,6 +7,7 @@
 package com.datadog.android.utils.forge
 
 import com.datadog.android.v2.api.context.UserInfo
+import com.datadog.tools.unit.forge.exhaustiveAttributes
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
