+++ tools/unit/src/main/kotlin/com/datadog/tools/unit/forge/GsonJsonObjectForgeryFactory.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.utils.forge
+package com.datadog.tools.unit.forge
 
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
@@ -12,12 +12,16 @@ import com.google.gson.JsonPrimitive
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
 
+/**
+ *  A [ForgeryFactory] generating a random [JsonObject] instance.
+ */
 class GsonJsonObjectForgeryFactory : ForgeryFactory<JsonObject> {
 
+    /** @inheritDoc */
     override fun getForgery(forge: Forge): JsonObject {
         return JsonObject().apply {
             val fieldCount = forge.aTinyInt()
-            repeat(fieldCount) {
+            repeat(fieldCount + 1) {
                 add(
                     forge.anAlphabeticalString(),
                     forge.anElementFrom(
