+++ tools/unit/src/main/kotlin/com/datadog/tools/unit/forge/OrgJSONObjectForgeryFactory.kt
@@ -4,15 +4,19 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.utils.forge
+package com.datadog.tools.unit.forge
 
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryFactory
 import org.json.JSONArray
 import org.json.JSONObject
 
+/**
+ *  A [ForgeryFactory] generating a random [JSONObject] instance.
+ */
 class OrgJSONObjectForgeryFactory : ForgeryFactory<JSONObject> {
 
+    /** @inheritDoc */
     override fun getForgery(forge: Forge): JSONObject {
         return JSONObject().apply {
             val fieldCount = forge.aTinyInt()
